+++ src/nile/common.py
@@ -5,6 +5,10 @@ import os
 import re
 import subprocess
 
+from starkware.crypto.signature.fast_pedersen_hash import pedersen_hash
+from starkware.starknet.core.os.class_hash import compute_class_hash
+from starkware.starknet.services.api.contract_class import ContractClass
+
 from nile.utils import normalize_number, str_to_felt
 
 CONTRACTS_DIRECTORY = "contracts"
@@ -170,3 +174,20 @@ def is_string(param):
 def is_alias(param):
     """Identify param as alias (instead of address)."""
     return is_string(param)
+
+
+def get_contract_class(contract_name, overriding_path=None):
+    """Return the contract_class for a given contract name."""
+    base_path = (
+        overriding_path if overriding_path else (BUILD_DIRECTORY, ABIS_DIRECTORY)
+    )
+    with open(f"{base_path[0]}/{contract_name}.json", "r") as fp:
+        contract_class = ContractClass.loads(fp.read())
+
+    return contract_class
+
+
+def get_hash(contract_name, overriding_path=None):
+    """Return the class_hash for a given contract name."""
+    contract_class = get_contract_class(contract_name, overriding_path)
+    return compute_class_hash(contract_class=contract_class, hash_func=pedersen_hash)
